pyproject.toml
[tool.poetry]
name = "flask_prj"
version = "0.1.0"
description = ""
authors = ["KosarevPV <pavel999kosarev@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
flask = "^2.2.3"
Flask-SQLAlchemy = "^3.0.3"
poetry-plugin-dotenv = "^0.4.1"
python-dotenv = "^1.0.0"
flask-login = "0.5.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
